# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created"
          }
        },
        "summary": "Create a post (content calendar entry)",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "review",
                "approved",
                "scheduled",
                "published",
                "analyzing",
                "underperforming",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List posts (optionally by campaign/status) — content calendar feed",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/ads": {
      "post": {
        "operationId": "AdsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ad created"
          }
        },
        "summary": "Create an ad (with derived CTR/CPC/CPA)",
        "tags": [
          "ads"
        ]
      },
      "get": {
        "operationId": "AdsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List ads (optionally by campaign)",
        "tags": [
          "ads"
        ]
      }
    },
    "/api/ads/{id}": {
      "get": {
        "operationId": "AdsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get an ad with derived metrics",
        "tags": [
          "ads"
        ]
      },
      "patch": {
        "operationId": "AdsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update an ad",
        "tags": [
          "ads"
        ]
      }
    },
    "/api/ads/{id}/analyze": {
      "post": {
        "operationId": "AdsController_analyze",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "AI optimization analysis for the ad’s campaign",
        "tags": [
          "ads"
        ]
      }
    },
...
    "/api/approvals": {
      "get": {
        "operationId": "ApprovalsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected",
                "edited"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List approval requests",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{id}": {
      "get": {
        "operationId": "ApprovalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get an approval request",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{id}/approve": {
      "post": {
        "operationId": "ApprovalsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Approve and execute the gated action",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{id}/reject": {
      "post": {
        "operationId": "ApprovalsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reject (captured as a learning example)",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{id}/edit": {
      "post": {
        "operationId": "ApprovalsController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Edit + approve (captures the human-edit delta as learning)",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/posts/{postId}/audience": {
      "get": {
        "operationId": "AudienceController_get",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Audience comments + the engagement agent’s latest summary",
        "tags": [
          "audience"
        ]
      }
    },
    "/api/posts/{postId}/audience/{commentId}/reply": {
      "post": {
        "operationId": "AudienceController_reply",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Send a reply to an audience comment (mock) + record it in memory",
        "tags": [
          "audience"
        ]
      }
    },
...
    "/api/agents/run": {
      "post": {
        "operationId": "AgentsController_run",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Completed agent run with steps"
          }
        },
        "summary": "Run an agent and return its run with timeline",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/runs": {
      "get": {
        "operationId": "AgentsController_listRuns",
        "parameters": [
          {
            "name": "agentType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "missionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List agent runs",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/runs/{id}": {
      "get": {
        "operationId": "AgentsController_getRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get an agent run with full timeline + reflections",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/posts/{id}/simulate": {
      "post": {
        "operationId": "PostActionsController_simulate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Swarm agent run with persona timeline"
          }
        },
        "summary": "Run the social simulation swarm on a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}/analyze": {
      "post": {
        "operationId": "PostActionsController_analyze",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Run the performance monitoring agent on a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}/approve": {
      "post": {
        "operationId": "PostActionsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Human-approve a post (moves it to approved)",
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hiive Agentic Marketing OS",
    "description": "Mission-driven, multi-agent marketing operating system: RAG memory, campaign intelligence, swarm simulation, autonomous monitoring, human-in-the-loop approvals, and a continuous learning flywheel.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "example": "campaign id"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "copy": {
            "type": "string",
            "example": "Thinking about liquidity before your IPO? ..."
          },
          "mediaUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "approved",
              "scheduled",
              "published",
              "analyzing",
              "underperforming",
              "completed"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-06-10T09:00:00.000Z"
          }
        },
        "required": [
          "campaignId",
          "platform",
          "copy"
        ]
      },
...
      "CreateAdDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Founder liquidity — value prop A"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "budget": {
            "type": "number",
            "example": 5000
          },
          "spend": {
            "type": "number",
            "example": 1200
          },
          "impressions": {
            "type": "number",
            "example": 40000
          },
          "clicks": {
            "type": "number",
            "example": 380
          },
          "conversions": {
            "type": "number",
            "example": 14
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "paused",
              "completed"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-06-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-06-30T00:00:00.000Z"
          }
        },
        "required": [
          "campaignId",
          "name",
          "platform",
          "budget"
        ]
      },
      "UpdateAdDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Founder liquidity — value prop A"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "budget": {
            "type": "number",
            "example": 5000
          },
          "spend": {
            "type": "number",
            "example": 1200
          },
          "impressions": {
            "type": "number",
            "example": 40000
          },
          "clicks": {
            "type": "number",
            "example": 380
          },
          "conversions": {
            "type": "number",
            "example": 14
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "paused",
              "completed"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-06-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-06-30T00:00:00.000Z"
          }
        }
      },
...
      "RunAgentDto": {
        "type": "object",
        "properties": {
          "agentType": {
            "type": "string",
            "example": "ContentGenerationAgent",
            "description": "Agent type to run"
          },
          "input": {
            "type": "object",
            "example": {
              "campaignId": "abc",
              "count": 3
            },
            "description": "Agent input payload"
          },
          "missionId": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          }
        },
        "required": [
          "agentType",
          "input"
        ]
      },
...
    }
  }
}